requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.2.0
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.0.1
PyMySQL==1.0.2
PyQt5==5.15.2
pyqt5-plugins==5.15.2.2.0.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqt5-tools==5.15.2.3.0.2
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
qt5-applications==5.15.2.2.1
qt5-tools==5.15.2.1.0.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
virtualenv==20.4.2
whitenoise==5.2.0
